It compiles

+++ components/Compute_Function/include/Compute_Function.h
@@ -1,1 +1,1 @@
-extern float Compute_Function(float x, int p, int q, int r);
+extern int Compute_Function(int x, int p, int q, int r);
