+++ components/Compute_Integral/include/Compute_Integral.h
@@ -1,1 +1,1 @@
-extern float Compute_Integral(int a, int b, int p, int q, int r, int N);
+extern int Compute_Integral(int a, int b, int p, int q, int r, int N);
